+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `Assent.Strategy.OAuth2.get_access_token/3` added
 * `Assent.Strategy.OAuth2.refresh_access_token/3` added
+* `Assent.Strategy.OAuth2.authorization_headers/2` is no long a public function
 
 ## v0.1.13 (2020-07-14)